+++ src/Tidier.jl
@@ -7,16 +7,17 @@ using Statistics
 using Cleaner
 using Reexport
 
-# Exporting `Cols` because `summarize(across(!!vars, funs))` with multiple interpolated
+# Exporting `Cols` because `summarize(!!vars, funs))` with multiple interpolated
 # columns requires `Cols()` to be nested within `Cols()`, so `Cols` needs to be exported.
 @reexport using DataFrames: DataFrame, Cols, describe, nrow, proprow
 @reexport using Chain
 @reexport using Statistics
+@reexport using ShiftedArrays: lag, lead
 
 export Tidier_set, across, desc, n, row_number, starts_with, ends_with, matches, if_else, case_when, ntile, 
       @select, @transmute, @rename, @mutate, @summarize, @summarise, @filter, @group_by, @ungroup, @slice, 
       @arrange, @distinct, @pull, @left_join, @right_join, @inner_join, @full_join, @pivot_wider, @pivot_longer, 
-      @bind_rows, @bind_cols, @clean_names, @count, @tally
+      @bind_rows, @bind_cols, @clean_names, @count, @tally, @drop_na
 
 # Package global variables
 const code = Ref{Bool}(false) # output DataFrames.jl code?
@@ -609,6 +610,48 @@ macro pull(df, column)
     @info MacroTools.prettify(vec_expr)
   end
   return vec_expr
+end
+
+"""
+$docstring_drop_na
+"""
+macro drop_na(df, exprs...)
+  interpolated_exprs = parse_interpolation.(exprs)
+
+  tidy_exprs = [i[1] for i in interpolated_exprs]
+
+  tidy_exprs = parse_tidy.(tidy_exprs)
+  num_exprs = length(exprs)
+  df_expr = quote
+    if $(esc(df)) isa GroupedDataFrame
+      local col_names = groupcols($(esc(df)))
+      @chain $(esc(df)) begin
+        DataFrame # remove grouping because `dropmissing()` does not work on GroupDataFrames
+        @chain _ begin
+          if $num_exprs == 0
+            dropmissing(_)
+          else
+            dropmissing(_, Cols($(tidy_exprs...)))
+          end
+        end
+        groupby(col_names; sort = true) # regroup
+      end
+    else
+      @chain $(esc(df)) begin
+        @chain _ begin
+          if $num_exprs == 0
+            dropmissing(_)
+          else
+            dropmissing(_, Cols($(tidy_exprs...)))
+          end
+        end
+      end
+    end
+  end
+  if code[]
+    @info MacroTools.prettify(df_expr)
+  end
+  return df_expr
 end
 
 end