+++ src/parsing.jl
@@ -290,9 +290,9 @@ function parse_autovec(tidy_expr::Union{Expr,Symbol})
       return :((∉).($arg1, Ref(Set($arg2))))
 
     elseif @capture(x, fn_(args__))
-
+      # This is the do-not-vectorize "list"
       # `in` should be vectorized so do not add to this exclusion list
-      if fn in [:Ref :Set :Cols :(:) :∘ :ntile :repeat :across :desc :mean :std :var :median :first :last :minimum :maximum :sum :length :skipmissing :quantile :passmissing]
+      if fn in [:Ref :Set :Cols :(:) :∘ :lag :lead :ntile :repeat :across :desc :mean :std :var :median :first :last :minimum :maximum :sum :length :skipmissing :quantile :passmissing]
         return x
       elseif contains(string(fn), r"[^\W0-9]\w*$") # valid variable name
         return :($fn.($(args...)))